+++ packages/firebase_auth/android/src/main/java/io/flutter/plugins/firebaseauth/FirebaseAuthPlugin.java
@@ -123,6 +123,9 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       case "signInWithPhoneNumber":
         handleSignInWithPhoneNumber(call, result);
         break;
+      case "setLanguageCode":
+        handleSetLanguageCode(call, result);
+        break;
       default:
         result.notImplemented();
         break;
@@ -151,9 +154,19 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         new PhoneAuthProvider.OnVerificationStateChangedCallbacks() {
           @Override
           public void onVerificationCompleted(PhoneAuthCredential phoneAuthCredential) {
-            Map<String, Object> arguments = new HashMap<>();
-            arguments.put("handle", handle);
-            channel.invokeMethod("phoneVerificationCompleted", arguments);
+            firebaseAuth
+                .signInWithCredential(phoneAuthCredential)
+                .addOnCompleteListener(
+                    new OnCompleteListener<AuthResult>() {
+                      @Override
+                      public void onComplete(@NonNull Task<AuthResult> task) {
+                        if (task.isSuccessful()) {
+                          Map<String, Object> arguments = new HashMap<>();
+                          arguments.put("handle", handle);
+                          channel.invokeMethod("phoneVerificationCompleted", arguments);
+                        }
+                      }
+                    });
           }
 
           @Override
@@ -464,6 +477,15 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
           String.format("Listener with identifier '%d' not found.", id),
           null);
     }
+  }
+
+  private void handleSetLanguageCode(MethodCall call, final Result result) {
+    @SuppressWarnings("unchecked")
+    Map<String, String> arguments = (Map<String, String>) call.arguments;
+    String language = arguments.get("language");
+
+    firebaseAuth.setLanguageCode(language);
+    result.success(null);
   }
 
   private class SignInCompleteListener implements OnCompleteListener<AuthResult> {